# src/static/swagger.json
{
    "info": {
        "description": "Auth service",
        "title": "Auth service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "summary": "\u0420\u0435\u0433\u0438\u0441\u0442\u0440\u0430\u0446\u0438\u044f \u043d\u043e\u0432\u043e\u0433\u043e \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Register"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New user was registered"
                    },
                    "409": {
                        "description": "Email is already in use"
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "summary": "\u0412\u0445\u043e\u0434 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f \u0432 \u0430\u043a\u043a\u0430\u0443\u043d\u0442",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Login"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "400": {
                        "description": "Wrong password"
                    },
                    "401": {
                        "description": "User is not exist"
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/change-password/{user_id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0421\u043c\u0435\u043d\u0430 \u043f\u0430\u0440\u043e\u043b\u044f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successful"
                    },
                    "400": {
                        "description": "Wrong password"
                    },
                    "401": {
                        "description": "User is not exist"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/refresh-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u041e\u0431\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u0442\u043e\u043a\u0435\u043d\u043e\u0432",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshToken"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refresh successful"
                    },
                    "400": {
                        "description": "Not user"
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0412\u044b\u0445\u043e\u0434 \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f \u0438\u0437 \u0430\u043a\u043a\u0430\u0443\u043d\u0442\u0430",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Logout"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged out"
                    },
                    "400": {
                        "description": "Not user"
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/login-history/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u041f\u043e\u043b\u0443\u0447\u0438\u0442\u044c \u0438\u0441\u0442\u043e\u0440\u0438\u044e \u043f\u043e\u0441\u0435\u0449\u0435\u043d\u0438\u0439",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged out"
                    },
                    "400": {
                        "description": "Not user"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
...
    },
...
    "openapi": "3.0.2",
    "components": {
        "schemas": {
            "Register": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "name",
                        "example": "TestUser"
                    },
                    "email": {
                        "type": "string",
                        "description": "email",
                        "example": "test_2@test.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "password",
                        "example": "Test_12345"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "username"
                ]
            },
            "Login": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "email",
                        "example": "test_2@test.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "password",
                        "example": "Test_12345"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "ChangePassword": {
                "type": "object",
                "properties": {
                    "old_password": {
                        "type": "string",
                        "description": "old_password",
                        "example": "Test_12345"
                    },
                    "new_password": {
                        "type": "string",
                        "description": "new_password",
                        "example": "Test_67890"
                    }
                },
                "required": [
                    "new_password",
                    "old_password"
                ]
            },
            "Tokens": {
                "type": "object",
                "properties": {
                    "refresh_token": {
                        "type": "string",
                        "description": "refresh_token",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTY2MzEyNDIzMiwianRpIjoiYzI2MzBlZmQtMjU1Ni00NDdiLTk1OGItMzZlOTIxNjg3MjM1IiwidHlwZSI6InJlZnJlc2giLCJzdWIiOiJhMjliMTI4ZC0zOWM3LTRjMTUtYmRhMi0wMjc4MzM5YjczMTMiLCJuYmYiOjE2NjMxMjQyMzIsImV4cCI6MTY2Mzk4ODIzMiwicGVybWlzc2lvbnMiOltdLCJpc19zdXBlciI6ZmFsc2V9.5Wmo7M7klrxO4_c47JI2xWAWSGljCS9cvn3D0VWUXmo"
                    },
                    "access_token": {
                        "type": "string",
                        "description": "access_token",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTY2MzEyNDIzMiwianRpIjoiM2UxOWQ2NmEtYTU4ZS00OGYzLWIyMTMtOTViYmZjYjgxOGY1IiwidHlwZSI6ImFjY2VzcyIsInN1YiI6ImEyOWIxMjhkLTM5YzctNGMxNS1iZGEyLTAyNzgzMzliNzMxMyIsIm5iZiI6MTY2MzEyNDIzMiwiZXhwIjoxNjYzMTI1NDMyLCJwZXJtaXNzaW9ucyI6W10sImlzX3N1cGVyIjpmYWxzZX0.Y0VDqXj8MtD2XlVJKG8nBS_HjQoJHL9oPcbjBvmZ2_k"
                    }
                },
                "required": [
                    "access_token",
                    "refresh_token"
                ]
            },
            "RefreshToken": {
                "type": "object",
                "properties": {
                    "tokens": {
                        "$ref": "#/components/schemas/Tokens"
                    }
                }
            },
            "Logout": {
                "type": "object",
                "properties": {
                    "from_all": {
                        "type": "boolean",
                        "description": "\u0412\u044b\u0445\u043e\u0434 \u0441\u043e \u0432\u0441\u0435\u0445 \u0443\u0441\u0442\u0440\u043e\u0439\u0441\u0442\u0432"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}